Implement bump_allocator and add tests
bump_allocator is a fast region based memory allocator.

We maintain three pointers into a backing buffer. One to the start, one
to the end, and one that tracks available space. To begin with the
tracking pointer points to the end of the backing buffer.

On allocation, we allocate downwards, from higher addresses to lower
addresses. We do a quick bounds check, then bump the tracking pointer
down.  If the current backing buffer is exhausted, we add it to a list
of used buffers and replace it with a new larger buffer.

On deallocation, all used buffers are freed and the tracking pointer is
reset to the end of the current backing buffer.

If AddressSanitizer is enabled we poison and unpoison memory as needed.
We also guard each allocation with a small poisoned memory region.