Implement process spawning utilities and add tests
Adds two utility classes, file_descriptor for managing the ownership of
a file descriptor, and unnamed_pipe for managing Unix pipes.

Implements a process creation API on top of the POSIX spawn functions.
The API allows the user to set the arguments, environment variables and
standard streams of the child process in a single function call.