Combine msg::int64 and msg::uint64 into one class
Previously msgpack integers were mapped into separate C++ objects
depending on their sign. This caused problems when type checking msgpack
objects - checking if an object contained an integer required two tests,
one for msg::int64 and one for msg::uint64.

Adds msg::integer class. This class represents a 64bit integer value, it
can be cast into signed an unsigned representations as required.

Sign information is lost when unpacking. We could add a boolean to
preserve signedness, but it seems pointless. We usually know if we
require a signed or unsigned representation at the point of use.