Rework undercurl drawing
Previously we applied a periodic function over the current x pixel
coordinate to produce a dotted line, this didn't guarantee that each
line began with a full length dash. This change attempts to fix this.

Track each cells position in an undercurl, and apply a periodic function
over the current pixel position in the line.