Make map_view an unsorted array of key value pairs
Map views were sorted to support reasonably fast lookup on large maps.
It turns out, for our use case, most maps are small and lookup is not
used very often. Map views are now an unsorted array of key value pairs,
with get() implemented as a linear search. This allows us to save on an
array sort.