Improve window scrolling when dragging with mouse
Continuously send mouse drag events to nvim (regardless of whether the
mouse has moved or not). This is required because when the mouse is
dragged beyond the bounds of a nvim window it should scroll the window
even when the mouse is stationary.

Introduce a timer based approach to keep nvim updated on mouse drag
events.

Note: This commit requires the following PR to be merged into nvim:
https://github.com/neovim/neovim/pull/22798